Added Packer debug option.
If the user chooses, the Packer script execution is held off so the user can log in to the Packer container and run the DebugPacker.sh script.  This runs Packer in debug mode, allowing the user to step through all the Packer activities interactively.

Also corrected region example in script update config.